requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-telegram-bot
google-generativeai
python-dotenv
pytest
httpx
